Enable users to update the script on the go.
Since most of the time the script needs to be updated, I have added the update script in the same shell file so that users can save a little bit of their time.
If an update is available, uses will just press 'y'. Else the program would execute as it is.